BDOG-473 small refactorings and update readme
